Fix segfault
in Window::render(), the semaphores were gotten with a structure binding which created a stack-allocated pointer to the semaphore. When returning the submit info, the semephores were destroyed as the stack was freed. The semaphores are now taken by reference so the pointer will always point to the heap where they wont be freed